Android - Return an error object from downloadFile
This is to try and understand what the cause of some
errors on certain devices (Android 10) which I cannot
replicate.

Only under certain error conditions. Note it can still
return a plain integer error code, so the result needs
to be checked to determine what format it is in.

If returning the object include properties:
	'code'  - The error code
	'error' - The exception as a string

If the url starts with https://dev then include the
stacktrace in the property 'trace'

Also if we have a file then include the files absolute
path in the 'file' property.